Fix build failures
