+++ performance/midi/collectnoteevents.go
@@ -2,6 +2,7 @@ package midi
 
 import (
 	"fmt"
+	"math"
 	"math/big"
 
 	"github.com/jamestunnell/go-musicality/performance/function"
@@ -41,20 +42,8 @@ func CollectNoteEvents(fs *model.FlatScore, dc *model.Computer, part string) ([]
 			return []*Event{}, err
 		}
 
-		vel, err := Velocity(n.Attack * dynamic)
-		if err != nil {
-			err = fmt.Errorf("failed to get MIDI velocity: %w", err)
-
-			return []*Event{}, err
-		}
-
-		newDur, err := model.AdjustDuration(pd.Duration, n.Separation)
-		if err != nil {
-			err = fmt.Errorf("failed to adjust duration: %w", err)
-
-			return []*Event{}, err
-		}
-
+		vel := Velocity(n.Attack * dynamic)
+		newDur := model.AdjustDuration(pd.Duration, n.Separation)
 		endOffset := new(big.Rat).Add(n.Start, newDur)
 
 		events = append(events, NewNoteOnEvent(n.Start, key, vel))
@@ -62,4 +51,29 @@ func CollectNoteEvents(fs *model.FlatScore, dc *model.Computer, part string) ([]
 	}
 
 	return events, nil
+}
+
+// Key converts the pitch to a MIDI note number.
+// Returns a non-nil error if the pitch is not in the range [C-1, G9].
+func Key(p *model.Pitch) (uint8, error) {
+	const (
+		// minTotalSemitone is the total semitone value of MIDI note 0 (octave below C0)
+		minTotalSemitone = -12
+		// maxTotalSemitone is the total semitone value of MIDI note 127 (G9)
+		maxTotalSemitone = 115
+	)
+
+	totalSemitone := p.TotalSemitone()
+
+	if totalSemitone < minTotalSemitone || totalSemitone > maxTotalSemitone {
+		return 0, fmt.Errorf("pitch %s is outside of MIDI note number range", p.String())
+	}
+
+	return uint8(totalSemitone + 12), nil
+}
+
+// Velocity converts the attack to a MIDI velocity.
+// Returns a non-nil error if the attack is not in the range [0.0, 1.0]
+func Velocity(attack float64) uint8 {
+	return 31 + uint8(math.Round(attack*96))
 }
