+++ performance/model/adjustduration.go
@@ -4,30 +4,22 @@ import (
 	"math/big"
 
 	"github.com/jamestunnell/go-musicality/notation/note"
-	"github.com/jamestunnell/go-musicality/validation"
 )
 
-func AdjustDuration(dur *big.Rat, separation float64) (*big.Rat, error) {
-	err := validation.VerifyInRangeFloat(
-		"separation", separation, note.SeparationMin, note.SeparationMax)
-	if err != nil {
-		return nil, err
-	}
-
+func AdjustDuration(dur *big.Rat, separation float64) *big.Rat {
 	var adjust *big.Rat
 	switch separation {
-	case note.SeparationMin:
+	case note.ControlMin:
 		adjust = big.NewRat(1, 1)
-	case note.SeparationMax:
+	case note.ControlMax:
 		adjust = big.NewRat(1, 8)
 	default:
-		mul := (separation * 0.5) + 0.5
-		remove := new(big.Rat).Mul(new(big.Rat).SetFloat64(mul), big.NewRat(7, 8))
+		remove := new(big.Rat).Mul(new(big.Rat).SetFloat64(separation), big.NewRat(7, 8))
 		adjust = new(big.Rat).Sub(big.NewRat(1, 1), remove)
 	}
 
 	newDur := new(big.Rat).Mul(dur, adjust)
 
-	return newDur, nil
+	return newDur
 
 }
